+++ services/backend/src/database/migrations/2021.05.20-1543-addIpAddressAllowListTable.js
@@ -1,0 +1,27 @@
+module.exports = {
+  up: async (queryInterface, Sequelize) => {
+    await queryInterface.createTable('IPAddressAllowList', {
+      ip: {
+        type: Sequelize.INET,
+        allowNull: false,
+        primaryKey: true,
+      },
+      comment: {
+        type: Sequelize.STRING,
+        allowNull: true,
+      },
+      rateLimitFactor: {
+        type: Sequelize.INTEGER,
+        allowNull: true,
+      },
+    });
+    await queryInterface.addIndex('IPAddressAllowList', {
+      fields: [{ name: 'ip', operator: 'inet_ops' }],
+      using: 'gist',
+    });
+  },
+
+  down: async queryInterface => {
+    await queryInterface.dropTable('IPAddressAllowList');
+  },
+};